VectorizationBase.jl

VectorizationBase.AbstractStridedPointer

VectorizationBase.GroupedStridedPointers

VectorizationBase.MM

VectorizationBase.Unroll

VectorizationBase._vrangeincr

VectorizationBase.align

VectorizationBase.bitselect

VectorizationBase.grouped_strided_pointer

VectorizationBase.ifmahi

VectorizationBase.ifmalo

VectorizationBase.inv_approx

VectorizationBase.lazymul

VectorizationBase.offset_ptr

VectorizationBase.pause

VectorizationBase.preserve_buffer

VectorizationBase.promote_div

VectorizationBase.unrolled_indicies

VectorizationBase.AbstractStridedPointer — Type

abstract type AbstractStridedPointer{T,N,C,B,R,X,O} end

T: element type N: dimensionality C: contiguous dim B: batch size R: rank of strides X: strides O: offsets

VectorizationBase.GroupedStridedPointers — Type

P are the pointers I contains indexes into strides (dynamic) and X (static) X static strides

VectorizationBase.MM — Type

The name MM type refers to MM registers such as XMM, YMM, and ZMM. MMX from the original MMX SIMD instruction set is a [meaningless initialism](https://en.wikipedia.org/wiki/MMX(instruction_set)#Naming).

The MM{W,X} type is used to represent SIMD indexes of width W with stride X.

VectorizationBase.Unroll — Type

Unroll{AU,F,N,AV,W,M,X}(i::I)

AU: Unrolled axis

F: Factor, step size per unroll. If AU == AV, F == W means successive loads. 1 would mean offset by 1, e.g. x{1:8], x[2:9], and x[3:10].

N: How many times is it unrolled

AV: Vectorized axis # 0 means not vectorized, some sort of reduction

W: vector width

M: bitmask indicating whether each factor is masked

X: stride between loads of vectors along axis AV.

i::I - index

VectorizationBase._vrangeincr — Method

vrange(::Val{W}, i::I, ::Val{O}, ::Val{F})

W - Vector width i::I - dynamic offset O - static offset F - static multiplicative factor

VectorizationBase.align — Function

align(x::Union{Int,Ptr}, [n])

Return aligned memory address with minimum increment. align assumes n is a power of 2.

VectorizationBase.bitselect — Method

bitselect(m::Unsigned, x::Unsigned, y::Unsigned)

If you have AVX512, setbits of vector-arguments will select bits according to mask m, selecting from x if 0 and from y if 1. For scalar arguments, or vector arguments without AVX512, setbits requires the additional restrictions on y that all bits for which m is 1, y must be 0. That is for scalar arguments or vector arguments without AVX512, it requires the restriction that ((y ⊻ m) & m) == m

VectorizationBase.grouped_strided_pointer — Method

G is a tuple(tuple((Aind,A's dim),(Aind,A's dim)), ()) it gives the groups.

VectorizationBase.ifmahi — Method

ifmalo(v1, v2, v3)

Multiply unsigned integers v1 and v2, adding the upper 52 bits to v3.

Requires has_feature(Val(:x86_64_avx512ifma)) to be fast.

VectorizationBase.ifmalo — Method

ifmalo(v1, v2, v3)

Multiply unsigned integers v1 and v2, adding the lower 52 bits to v3.

Requires has_feature(Val(:x86_64_avx512ifma)) to be fast.

VectorizationBase.inv_approx — Method

Fast approximate reciprocal.

Guaranteed accurate to at least 2^-14 ≈ 6.103515625e-5.

Useful for special funcion implementations.

VectorizationBase.lazymul — Method

Basically:

if I ∈ [3,5,7,9] c[(I - 1) >> 1] else b * I end

because

c = b .* [3, 5, 7, 9]

VectorizationBase.offset_ptr — Method

An omnibus offset constructor.

The general motivation for generating the memory addresses as LLVM IR rather than combining multiple lllvmcall Julia functions is that we want to minimize the inttoptr and ptrtoint calculations as we go back and fourth. These can get in the way of some optimizations, such as memory address calculations. It is particulary import for gather and scatters, as these functions take a Vec{W,Ptr{T}} argument to load/store a Vec{W,T} to/from. If sizeof(T) < sizeof(Int), converting the <W x $(typ)* vectors of pointers in LLVM to integer vectors as they're represented in Julia will likely make them too large to fit in a single register, splitting the operation into multiple operations, forcing a corresponding split of the Vec{W,T} vector as well. This would all be avoided by not promoting/widenting the <W x $(typ)> into a vector of Ints.

For this last issue, an alternate workaround would be to wrap a Vec of 32-bit integers with a type that defines it as a pointer for use with internal llvmcall functions, but I haven't really explored this optimization.

VectorizationBase.pause — Method

pause()

For use in spin-and-wait loops, like spinlocks.

VectorizationBase.preserve_buffer — Method

For structs wrapping arrays, using GC.@preserve can trigger heap allocations. preserve_buffer attempts to extract the heap-allocated part. Isolating it by itself will often allow the heap allocations to be elided. For example:

julia> using StaticArrays, BenchmarkTools

julia> # Needed until a release is made featuring https://github.com/JuliaArrays/StaticArrays.jl/commit/a0179213b741c0feebd2fc6a1101a7358a90caed
       Base.elsize(::Type{<:MArray{S,T}}) where {S,T} = sizeof(T)

julia> @noinline foo(A) = unsafe_load(A,1)
foo (generic function with 1 method)

julia> function alloc_test_1()
           A = view(MMatrix{8,8,Float64}(undef), 2:5, 3:7)
           A[begin] = 4
           GC.@preserve A foo(pointer(A))
       end
alloc_test_1 (generic function with 1 method)

julia> function alloc_test_2()
           A = view(MMatrix{8,8,Float64}(undef), 2:5, 3:7)
           A[begin] = 4
           pb = parent(A) # or `LoopVectorization.preserve_buffer(A)`; `perserve_buffer(::SubArray)` calls `parent`
           GC.@preserve pb foo(pointer(A))
       end
alloc_test_2 (generic function with 1 method)

julia> @benchmark alloc_test_1()
BenchmarkTools.Trial:
  memory estimate:  544 bytes
  allocs estimate:  1
  --------------
  minimum time:     17.227 ns (0.00% GC)
  median time:      21.352 ns (0.00% GC)
  mean time:        26.151 ns (13.33% GC)
  maximum time:     571.130 ns (78.53% GC)
  --------------
  samples:          10000
  evals/sample:     998

julia> @benchmark alloc_test_2()
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     3.275 ns (0.00% GC)
  median time:      3.493 ns (0.00% GC)
  mean time:        3.491 ns (0.00% GC)
  maximum time:     4.998 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000

VectorizationBase.promote_div — Method

Promote, favoring <:Signed or <:Unsigned of first arg.

VectorizationBase.unrolled_indicies — Method

Returns a vector of expressions for a set of unrolled indices.